Overdue commit - multiple stuffs
- historic2.py trying a more general approach whereby the history
is a dictionary of actual function calls, so we don't have to assume
an object with an .undo() method
- possibilities.py is a module providing a 'possibilities set' possSet
which pretty much follows the historic item protocol of historic.py
(without actually including it) while giving a streamlined implementation
of a set that:
  - can only have items removed (except by undo)
  - raises Contradiction exception when emptied
It is basically a separated off piece of...
- tryposit.py is work toward experimenting with try-except statements
as the positing action (hence above Contradiction as an exception).
Essentially we want to know if they behave properly under recursive nesting.
The test module is based on solbing n queens puzzle.