Lots of stuff - mostly on n queens
Some distracted investigations. queens3 was a branch from queens2 with
the intention of focussing more on that specific problem as distinct
from the try-posit and historic projects. In particular, implementing
extra restrictions to only return 'normal' solutions, and not their
rotations and reflections, which can be much more easily generated from
them. Each asymmetric 'normal' solution represents 8 solutions in the
full set. (4 for 2-way rotational symmetry, 2 for 4-way.) That hasn't
yet fully happened (there are some intricacies) but another diversion
has occurred in the form of nqueens3a which is a standalone n-queens
solver - no try-posit or historic architecture, and a non-recursive
main loop. It uses long integers as binary bit-maps for the whole board
(numpy arrays proved no quicker). This thinking spurred the development
of rangesubset.py - a much more structured bit of programming providing
a customised subclass of set with a bit-map back-end. nqueens3a uses the
same idea but without the module. (Everything direct for speed).
nqueens3a1.py is a backup of an earlier version of nqueens3a with
the main execution leap structured differently.